+++ releasenotes.md
@@ -13,6 +13,7 @@
 - Give some symbol name suggestions when the path is matched.
 - Don't generate .o files on `compile` and `compile-run` if there is no `main`.
 - c3c init-lib does not create the directory with the .c3l suffix #1253
+- Permit foreach values to be optional.
 
 ### Fixes
 - Broken WASM library code.